+++ eras/conway/impl/src/Cardano/Ledger/Conway/TxBody.hs
@@ -47,6 +47,8 @@ module Cardano.Ledger.Conway.TxBody (
   ConwayTxBodyRaw (..),
   conwayTotalDepositsTxBody,
   conwayProposalsDeposits,
+  conwayRedeemerPointer,
+  conwayRedeemerPointerInverse,
 ) where
 
 import Cardano.Ledger.Alonzo.TxBody (Indexable (..))
@@ -64,6 +66,7 @@ import Cardano.Ledger.Binary (
   Sized (..),
   ToCBOR (..),
   mkSized,
+  unsafeMapSized,
  )
 import Cardano.Ledger.Binary.Coders (
   Decode (..),
@@ -338,11 +341,7 @@ instance EraTxBody ConwayEra where
     pure $
       ConwayTxBody
         { ctbSpendInputs = btbInputs btb
-        , ctbOutputs =
-            mkSized (eraProtVerLow @ConwayEra)
-              . upgradeTxOut
-              . sizedValue
-              <$> btbOutputs btb
+        , ctbOutputs = unsafeMapSized upgradeTxOut <$> btbOutputs btb
         , ctbCerts = certsOSet
         , ctbWithdrawals = btbWithdrawals btb
         , ctbTxfee = btbTxFee btb
@@ -354,11 +353,7 @@ instance EraTxBody ConwayEra where
         , ctbScriptIntegrityHash = btbScriptIntegrityHash btb
         , ctbTxNetworkId = btbTxNetworkId btb
         , ctbReferenceInputs = btbReferenceInputs btb
-        , ctbCollateralReturn =
-            mkSized (eraProtVerLow @ConwayEra)
-              . upgradeTxOut
-              . sizedValue
-              <$> btbCollateralReturn btb
+        , ctbCollateralReturn = unsafeMapSized upgradeTxOut <$> btbCollateralReturn btb
         , ctbTotalCollateral = btbTotalCollateral btb
         , ctbCurrentTreasuryValue = SNothing
         , ctbProposalProcedures = OSet.empty