+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/UtxosSpec.hs
@@ -701,11 +701,11 @@ testPlutusV1V2Failure sh badField lenz errorField = do
 
 enactCostModels ::
   ConwayEraImp era =>
-  StrictMaybe (GovPurposeId 'PParamUpdatePurpose era) ->
+  StrictMaybe (GovPurposeId 'PParamUpdatePurpose) ->
   CostModels ->
   Credential 'DRepRole ->
   NonEmpty (Credential 'HotCommitteeRole) ->
-  ImpTestM era (GovPurposeId 'PParamUpdatePurpose era)
+  ImpTestM era (GovPurposeId 'PParamUpdatePurpose)
 enactCostModels prevGovId cms dRep committeeMembers' = do
   initialCms <- getsNES $ nesEsL . curPParamsEpochStateL . ppCostModelsL
   let pparamsUpdate = def & ppuCostModelsL .~ SJust cms